Added moving relative layout
I added class Rig which contains a layout that can be moved by holding
down it's child "move" button.   Still in the works but we must find a
way to attach the camera class that we already have to the rig.  Also,
while doing this I implemented new drag code and fixed the margins.